# Conflicts:
#	.idea/misc.xml
#	app/src/main/AndroidManifest.xml
#	app/src/main/java/com/wd/tech/Fragment/InformationFragment.kt
#	app/src/main/java/com/wd/tech/Fragment/MineFragment.kt
#	app/src/main/java/com/wd/tech/activity/IndividualInformationActivity.kt
#	app/src/main/java/com/wd/tech/activity/LoginActivity.kt
#	app/src/main/java/com/wd/tech/activity/MainActivity.kt
#	app/src/main/java/com/wd/tech/adapter/CommunityAdapter.kt
#	app/src/main/java/com/wd/tech/adapter/CommunityFileAdapter.kt
#	app/src/main/java/com/wd/tech/api/Api.kt
#	app/src/main/java/com/wd/tech/app/MyApp.kt
#	app/src/main/java/com/wd/tech/base/BaseActivity.kt
#	app/src/main/java/com/wd/tech/utils/JumpActivityUtils.kt

+++ app/src/main/java/com/wd/tech/bean/InfoDetailsBean.kt
@@ -24,7 +24,7 @@ data class InfoDetailsResult(
     var title: String,
     var whetherCollection: Int,
     var whetherGreat: Int,
-    var yuanCost: Int
+    var yuanCost: Double
 )
 
 data class Information(
